refactor: rename native object cleaner thread and native object phantom reference
