+++ src/main/java/org/rocstreaming/roctoolkit/NativeObjectCleaner.java
@@ -1,0 +1,88 @@
+package org.rocstreaming.roctoolkit;
+
+import java.lang.ref.ReferenceQueue;
+import java.util.Collections;
+import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
+
+/**
+ * Thread collecting references and closing {@link NativeObject}s when they become phantom reachable.
+ */
+class NativeObjectCleaner extends Thread {
+
+    /**
+     * Singleton instance.
+     */
+    private final static NativeObjectCleaner instance = new NativeObjectCleaner();
+
+    /**
+     * Queue of phantom reachable {@link NativeObject}.
+     */
+    private final ReferenceQueue<NativeObject> referenceQueue;
+
+    /**
+     * Set to keep references to prevent being garbage collected
+     */
+    private final Set<NativeObjectPhantomReference> set = Collections.newSetFromMap(new ConcurrentHashMap<>());
+
+    /**
+     * Create a new <code>NativeObjectCleaner</code>.
+     */
+    private NativeObjectCleaner() {
+        super("RocNativeObjectCleaner");
+        setDaemon(true);
+        referenceQueue = new ReferenceQueue<>();
+    }
+
+    /**
+     * Get <code>NativeObjectCleaner</code> instance.
+     *
+     * @return the <code>NativeObjectCleaner</code> singleton instance.
+     */
+    static NativeObjectCleaner getInstance() {
+        return instance;
+    }
+
+    /**
+     * Register a {@link NativeObject} in <code>NativeObjectCleaner</code>.
+     *
+     * @param nativeObj  {@link NativeObject} to add.
+     * @param ptr        Underlying roc object native pointer.
+     * @param destructor Destructor method.
+     * @return the new {@link NativeObjectPhantomReference} associated to the {@link NativeObject}.
+     */
+    NativeObjectPhantomReference register(NativeObject nativeObj, long ptr, Destructor destructor) {
+        NativeObjectPhantomReference reference = new NativeObjectPhantomReference(nativeObj, referenceQueue, ptr, destructor);
+        set.add(reference);
+        return reference;
+    }
+
+    /**
+     * Remove a reference from <code>NativeObjectCleaner</code>.
+     *
+     * @param reference     the {@link NativeObjectPhantomReference} to remove.
+     */
+    void remove(NativeObjectPhantomReference reference) {
+        set.remove(reference);
+    }
+
+    /**
+     * Entrypoint method of <code>NativeObjectCleaner</code>.
+     *
+     * Remove any phantom reachable {@link NativeObjectPhantomReference} from the
+     * {@link NativeObjectCleaner#referenceQueue} associated with this <code>NativeObjectCleaner</code>
+     * and <code>close</code> it.
+     */
+    @Override
+    public void run() {
+        while (isAlive()) {
+            try {
+                NativeObjectPhantomReference reference = (NativeObjectPhantomReference) referenceQueue.remove();
+                set.remove(reference);
+                reference.close();
+            } catch (Exception e) {
+            }
+        }
+    }
+
+}
